+++ docs/README.md
@@ -1,0 +1,12 @@
+- 로또 구매 (구매 금액 입력)
+  PurchasedLotto: 구매 개수, 발행된 로또
+
+- 로또 발행 (구매한 로또 출력)
+
+- 당첨 로또 (당첨 번호 입력)
+  WinningLotto: 당첨 번호, 보너스 번호
+
+- 당첨 통계 출력
+  WinningRank: 등수와 당첨금, enum 사용
+
+- Exception 처리
