+++ src/main/java/lotto/controller/LottoController.java
@@ -1,0 +1,26 @@
+package lotto.controller;
+
+public class LottoController {
+    public void run() {
+        purchaseLotto();
+        publishLotto();
+        getLottoNumbers();
+        printLottoResult();
+    }
+
+    private void purchaseLotto() {
+
+    }
+
+    private void publishLotto() {
+
+    }
+
+    private void getLottoNumbers() {
+
+    }
+
+    private void printLottoResult() {
+
+    }
+}
